requirements.txt
-i https://pypi.org/simple
bson==0.5.8
dataclasses==0.6
django-crispy-forms==1.8.1
django-templated-mail==1.1.1
django==2.2.10
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
djongo==1.3.1
djoser==2.0.3
mysqlclient==1.4.6
pillow==7.0.0
pyjwt==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
